Create Fractal-update
This program prints the output much more quickly, which is the most obvious improvement when interacting with it, but it also is much clearer about what is happening, it works entirely recursively instead of altering a list by using the pointer to it, and it creates the board rather than starting with a board of the right size and removing elements from it. Of course, the comments are also improved. This shows my current ability conceptually, rather than technically.